raspberry: Prevent overzealous buffer cycling in set_enabled_unlocked()
With Raspberry Pi the FillBufferDone callback is called *during* the
OMX_FillThisBuffer call, which appears to be allowed by the recursive
locking. Thus the buffer to be filled can already be again in the queue
after OMX_FillThisBuffer returns and the while loop will fill it again,
effectively discarding the old contents before handling them.